backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
psycopg2==2.9.6
requests==2.31.0
Werkzeug==2.2.3
gunicorn==20.1.0
